% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_data.R
\name{ajuste_data}
\alias{ajuste_data}
\title{Ajuste em colunas com formato de data}
\usage{
ajuste_data(df, tipo_data = 1)
}
\arguments{
\item{df}{Nome do data frame.}

\item{tipo_data}{Tipo do formato de data: \code{1} para datas no formato "DD/MM/YYYY"
ou \code{2} para datas no formato "YYYY-MM-DD".}
}
\value{
O data frame modificado, onde as colunas que comecam com "dt_" ou "DT_" foram convertidas
para o formato de data especificado pelo parametro \code{tipo_data}.
}
\description{
Converte automaticamente colunas que começam com "dt_" ou "DT_" em objetos do tipo \code{Date},
usando o formato especificado via parametro.
}
\examples{
df <- data.frame(dt_nasc = c("01/01/2020", "15/12/2021"))
ajuste_data(df, tipo_data = 1)

df2 <- data.frame(DT_OBITO = c("2022-05-10", "2023-01-01"))
ajuste_data(df2, tipo_data = 2)

}
