% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustes_nomes.R
\name{ajustes_nomes}
\alias{ajustes_nomes}
\title{Ajuste e Limpeza Padronizada de Nomes em Data Frames}
\usage{
ajustes_nomes(df)
}
\arguments{
\item{df}{Um data frame contendo colunas de nomes de pessoas a serem padronizadas.}
}
\value{
O data frame original com colunas textuais ajustadas.
}
\description{
A funcao \code{ajustes_nomes()} realiza a limpeza e padronizaçao de colunas textuais
em um data frame, transformando expressões padronizadas em \code{NA},
removendo acentos, sufixos, conectivos e palavras irrelevantes em nomes proprios.

Aplica transformacoes apenas nas colunas que são \code{character},
e, no caso de colunas com nomes que contenham "_nome", realiza uma limpeza semantica adicional.
}
\details{
A funcao executa:
\enumerate{
\item Substituicao de termos como "IGNORADO", "DESCONHECIDO", "NAO DECLARADO" por \code{NA}.
\item Padronização de acentos e caixa alta.
\item Limpeza especifica para colunas com "_nome", incluindo:
\itemize{
\item Remocao de palavras como "RN", "NATIMORTO", "FETO".
\item Remocao de sufixos como "FILHO", "JUNIOR", "NETO".
\item Eliminacao de conectivos como "DA", "DE", "DOS", "MC", "D\\u2019".
\item Substituicao de caracteres nao alfabeticos e multiplas letras repetidas.
}
}
}
\examples{
df <- data.frame(nome_paciente = c("Joao da Silva", "RN MARIA DE SOUZA", "NATIMORTO", "Ignorado"))
ajustes_nomes(df)

}
