% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_raca_sim.R
\name{ds_raca_sim}
\alias{ds_raca_sim}
\title{Decodifica a variavel de raca/cor no formato do SIM}
\usage{
ds_raca_sim(df)
}
\arguments{
\item{df}{Um data frame contendo uma coluna \code{cd_raca} com os codigos de raca/cor.}
}
\value{
O mesmo data frame com uma nova coluna \code{ds_raca}, contendo os valores decodificados.
}
\description{
A funcao \code{ds_raca_sim()} converte os codigos numericos da variavel \code{cd_raca}
em descricoes padronizadas conforme o formato utilizado nos dados do SIM.

Os codigos seguem a convencao:
\itemize{
\item "1" = Branca
\item "2" = Preta
\item "3" = Amarela
\item "4" = Parda
\item "5" = Indigena
\item Qualquer outro valor = Ignorada
}
}
\examples{
df <- data.frame(cd_raca = c("1", "2", "5", "9"))
ds_raca_sim(df)

}
