% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finaliza_pareamento.R
\name{finaliza_pareamento}
\alias{finaliza_pareamento}
\title{Montar a função de finalizar pareamento}
\usage{
finaliza_pareamento(df = 1)
}
\arguments{
\item{df}{ainda a definir, mas deve ser um data frame ou número que representa o último id de pareamento.}
}
\value{
numero adicionado + 1
}
\description{
Montar a função de finalizar pareamento
}
